% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_case_adj.R
\name{col_case_adj}
\alias{col_case_adj}
\title{adjust variable names}
\usage{
col_case_adj(data, to_case = "upper")
}
\arguments{
\item{data}{your extract}
}
\description{
at the moment episurv changes the variable names to camelCase.
to make it easy to sync up with various scripts,
this function adjusts the variable names
to upper or lower to suit your needs
}
\examples{
# no examples

}
\keyword{EpiSurv}
\keyword{colnames}
