% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{correction}
\alias{correction}
\title{Correction to the dataset}
\usage{
correction(given_cond, given_edit)
}
\arguments{
\item{given_cond}{a boolean expression. the conditions to meet e.g. epiSurvNumber == "##-######-XX" etc}

\item{given_edit}{a named vector of corrections e.g. c(age = "2 yrs",sex = "Male")}
}
\description{
a convenience function, to be used within spot_clean
passes an individual correction
}
\examples{
# no examples

}
\keyword{EpiSurv}
\keyword{Numbers}
