% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_conf_prob.R
\name{filter_conf_prob}
\alias{filter_conf_prob}
\title{Filter confirmed and probable cases in an extract}
\usage{
filter_conf_prob(data)
}
\arguments{
\item{data}{your extract}
}
\description{
I know this being criminally lazy, but it feels right to make this function
}
\examples{
# no examples
}
\keyword{EpiSurv}
\keyword{Status}
