% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ages.R
\name{group_ages}
\alias{group_ages}
\title{Count Yesses in selected columns.}
\usage{
group_ages(age_vctr, age_breaks, age_labels)
}
\arguments{
\item{age_vctr}{the age column}

\item{age_breaks}{the breaks}

\item{age_labels}{the labels}
}
\description{
This function takes an ag column, breaks and labels,
and returns the appropriate age groups
}
\examples{
# no examples
}
\keyword{Age}
\keyword{EpiSurv}
\keyword{Groups}
