% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_yes.R
\name{is_yes}
\alias{is_yes}
\title{Count Yesses in selected columns.}
\usage{
is_yes(..., operation = "any")
}
\arguments{
\item{...}{variables you want checked.}

\item{operation}{a string - one of the following:
"any" returns TRUE if at least 1 Yes present.
"all" returns TRUE only if all columns are Yes
"n" returns the number of Yesses in the row.}

\item{data}{your extract}
}
\description{
This function takes selected columns and checks if Yes is in each row.
Unfortunately at the moment it needs the first argument "." if used within mutate
}
\examples{
# no examples
}
\keyword{EpiSurv}
\keyword{Numbers}
