% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_episurv_csv.R
\name{read_episurv_csv}
\alias{read_episurv_csv}
\title{Read an EpiSurv csv.
This works with new and old extracts.
Currently only knows the common fields.
Disease specific ones you will have to add yourself.}
\usage{
read_episurv_csv(file, col_types = NULL, ...)
}
\arguments{
\item{file}{your extract}

\item{col_types}{extra column formats to add. pass as a cols() function}
}
\description{
Read an EpiSurv csv.
This works with new and old extracts.
Currently only knows the common fields.
Disease specific ones you will have to add yourself.
}
\examples{
# no examples
}
\keyword{EpiSurv}
\keyword{csv}
\keyword{formatting}
