% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_clean.R
\name{spot_clean}
\alias{spot_clean}
\title{Spot clean a dataset}
\usage{
spot_clean(data, adjustments)
}
\arguments{
\item{data}{your extract}

\item{adjustments}{a list of corrections. see ?correction}
}
\description{
takes a list of corrections
and applies them to the dataset
}
\examples{
# no examples
}
\keyword{EpiSurv}
\keyword{cleaning}
\keyword{data}
