% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_episurv_numbers.R
\name{trim_episurv_numbers}
\alias{trim_episurv_numbers}
\title{Trim EpiSurv Numbers}
\usage{
trim_episurv_numbers(data, ...)
}
\arguments{
\item{data}{your extract}

\item{...}{variables you want trimmed.}
}
\description{
You can pass your episurv numbers to this and it returns them trimmed i.e.
the suffix (-AK) is removed.
}
\examples{
# no examples
}
\keyword{EpiSurv}
\keyword{Numbers}
