% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_episurv_csv.R, R/read_episurv_csv_old.R
\name{read_episurv_csv}
\alias{read_episurv_csv}
\title{Read an EpiSurv csv.
This works with new and old extracts.
Currently only knows the common fields.
Disease specific ones you will have to add yourself.}
\usage{
read_episurv_csv(file, ...)

read_episurv_csv(file, ...)
}
\arguments{
\item{file}{the name of the extract.}

\item{...}{other arguments to pass to read.csv}

\item{col_types}{extra column formats to add. pass as a cols() function}
}
\description{
This function wraps read.csv() for reading in EpiSurv csv files, which are very fiddly.
}
\examples{
# no examples
# no examples

}
\keyword{EpiSurv}
\keyword{csv}
\keyword{formatting}
